Created controller for post

+++ app/Post.php
@@ -2,6 +2,8 @@
 
 namespace App;
 
+
+use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Model;
 use Cviebrock\EloquentSluggable\Sluggable;
@@ -14,17 +16,17 @@ class Post extends Model
     const IS_DRAFT = 0;
     const IS_PUBLIC = 1;
 
-    protected $fillable = ['title', 'content'];
+    protected $fillable = ['title', 'content', 'date'];
 
     public function category()
     {
-        return $this->hasOne(Category::class);
+        return $this->belongsTo(Category::class);
 
     }
 
     public function author()
     {
-        return $this->hasOne(User::class);
+        return $this->belongsTo(User::class, 'user_id');
     }
 
     public function tags()
@@ -64,8 +66,16 @@ class Post extends Model
 
     public function remove()
     {
-        Storage::delete('uploads/'.$this->image);
+        $this->removeImage();
         $this->delete();
+    }
+
+    public function removeImage()
+    {
+        if($this->image != null)
+        {
+            Storage::delete('uploads/'.$this->image);
+        }
     }
 
     public function uploadImage($image)
@@ -73,9 +83,9 @@ class Post extends Model
 
         if($image == null) {return;}
 
-        Storage::delete('uploads/'.$this->image);
+        $this->removeImage();
         $filename = str_random(10).'.'.$image->extension();
-        $image->saveAs('uploads', $filename);
+        $image->storeAs('uploads', $filename);
         $this->image = $filename;
         $this->save();
     }
@@ -150,6 +160,22 @@ class Post extends Model
         return '/uploads' . $this->image;
 
     }
+
+    public function setDateAttribute($value)
+    {
+        $date = Carbon::createFromFormat('d/m/y', $value)->format('Y-m-d');
+
+        $this->attributes['date'] = $date;
+    }
+
+    public function getCategoryTitle()
+    {
+        return ($this->category != null) ? $this->category->title : 'Нет категории';
+    }
     
+    public function getTagsTitles()
+    {
+        return (!$this->tags->isEmpty()) ? implode(', ', $this->tags->pluck('title')->all()): 'Нет тегов';
+    }
 
 }
